fix(feed): relocate ViewerAuthResult to lib/models to satisfy UI->service boundary
The check_ui_service_boundary.sh ratchet failed CI: pooled_age_restricted_retry.dart
(a lib/screens/ file) imported the new result type from lib/services/, which the
ratchet flags as a new UI->service-layer import. ViewerAuthResult is a plain DTO
(sealed class, no service dependencies), so move it to lib/models/ — which the
presentation layer is allowed to import — and repoint every importer. No behavior
change; its unit test moves to test/models/.
