+++ mobile/lib/models/viewer_auth_result.dart
