+++ mobile/test/screens/feed/pooled_age_restricted_retry_test.dart
@@ -12,11 +12,11 @@ import 'package:models/models.dart';
 import 'package:openvine/blocs/video_playback_status/video_playback_status_cubit.dart';
 import 'package:openvine/blocs/video_playback_status/video_playback_status_state.dart';
 import 'package:openvine/l10n/generated/app_localizations.dart';
+import 'package:openvine/models/viewer_auth_result.dart';
 import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/screens/feed/pooled_age_restricted_retry.dart';
 import 'package:openvine/services/age_verification_service.dart';
 import 'package:openvine/services/media_auth_interceptor.dart';
-import 'package:openvine/services/viewer_auth_result.dart';
 
 class _MockMediaAuthInterceptor extends Mock implements MediaAuthInterceptor {}
