+++ mobile/test/services/media_viewer_auth_service_test.dart
@@ -5,11 +5,11 @@ import 'package:fake_async/fake_async.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:nostr_sdk/nostr_sdk.dart';
+import 'package:openvine/models/viewer_auth_result.dart';
 import 'package:openvine/services/auth_service.dart';
 import 'package:openvine/services/media_viewer_auth_service.dart';
 import 'package:openvine/services/nip98_auth_service.dart';
 import 'package:openvine/services/nostr_identity.dart';
-import 'package:openvine/services/viewer_auth_result.dart';
 
 class MockAuthService extends Mock implements AuthService {}
